refactor: 서버, 클라이언트 컴포넌트 분리 작업
- 닉네임/로그인 버튼 dynamic import 해제하고 클라이언트 컴포넌트로 변경
- 게시판 컴포넌트 리팩토링
