+++ src/service/postService/usePagination.ts
@@ -1,11 +1,11 @@
-import { PaginationNumGroupProps } from "@/components/postBoard/molecules/PaginationNumGroup";
+import { PaginationNumGroupParameter } from "@/components/postBoard/molecules/PaginationNumGroup";
 import { paginationAtom } from "@/stateStore/postAtom";
 import { useEffect, useState } from "react";
 import { useSetRecoilState } from "recoil";
 
 const pagesPerGroup = 13;
 
-function usePagination({ currentPage, totalPages }: PaginationNumGroupProps) {
+function usePagination({ currentPage, totalPages }: PaginationNumGroupParameter) {
   const [pageGroupStart, setPageGroupStart] = useState(1);
   const setCurrentPage = useSetRecoilState(paginationAtom);
